requirements.txt
# Základní dependencies pro knihovny ETL pipeline

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0

# Excel processing
openpyxl>=3.1.0

# Database
psycopg2-binary>=2.9.0

# Airflow
apache-airflow>=2.7.0

# Development
pytest>=7.2.0